change ckb and fee amount from u128 to u64

+++ src/ckb/types.rs
@@ -819,7 +819,7 @@ impl TryFrom<molecule_cfn::TxAckRBF> for TxAckRBF {
 pub struct Shutdown {
     pub channel_id: Hash256,
     pub close_script: Script,
-    pub fee: u128,
+    pub fee: u64,
 }
 
 impl From<Shutdown> for molecule_cfn::Shutdown {
